Added config support for Wave #pragma command directives.
git-svn-id: http://svn.boost.org/svn/boost/trunk@31980 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/wave/util/insert_whitespace_detection.hpp
@@ -209,7 +209,7 @@ public:
     bool must_insert(boost::wave::token_id current, StringT const &value)
     {
         using namespace boost::wave;
-        switch (current) {
+        switch (static_cast<unsigned int>(current)) {
         case T_NONREPLACABLE_IDENTIFIER:
         case T_IDENTIFIER: 
             return impl::handle_identifier(prev, beforeprev, value);
